Re-calculating CNCH2CCH ion-neutral reactions
Previously, the ion-molecule reaction rate coefficients for CNCH2CCH, C4H3CN, and CNCHCCH were the same, which seems to come from a polarizability of 11.35 and dipole moment of 5.14 D. These new coefficients were calculated with a dipole moment of 3.61 and polarizability of 7.175 from Woon and Herbst (2009). The dipole moment is the experimental value originally from McNaughton et al. (1988), while the polarizability was calculated by Woon and Herbst.