Delete CH+C3H6 reactions on grains
Based on Fig. 7 in Loison & Bergeat 2009, it seems unlikely that C3H6+CH 
would take place on the grain surface, because (1) the calculations do 
not take any surface effect into account, which can impact the way the 
reactants sit on the surface and therefore the possible orientations for 
reactivity; (2) the C4H7 intermediates can dissipate their energy fairly 
quickly to grains and stabilize, leaving an isomer of C4H7 to be the 
product.