Commenting out high-T reactions
Commenting out all reactions where Tmin is greater than 10 K. Many of these reactions are "duplicate" entries for ion-molecule reactions because the Su-Chesnavich equation switches form at a species-dependent temperature. Since the NAUTILUS code does not exclude reactions based on temperature range, these reactions with two entries were being counted twice.

Non-duplicate reactions with Tmin > 10 were also commented out because of how the NAUTILUS code calculates temperature-dependent rate coefficients. If the Tmin for a reaction is greater than the temperature set in parameters.in, the rate coefficient is calculated at the Tmin, not the set temperature. So even if a gas temperature of 10 K is set for the simulation, if the Tmin for a reaction is 300 the rate coefficient will be calculated using 300 K, not 10. This leads to unrealistic rate coefficients for certain reactions.