Commenting out additional duplicates
Photodissociation of HCOOCH3 via external phtotons and cosmic-ray induced photons was duplicated in the model with different product channels. The versions left uncommented are consistent with the entries in KIDA. Additionally the cosmic-ray induced photodissociation of benzene was previously revised to have C6H5 + H as the primary products, however this old entry assuming 100% of products are C6H4 + H2 was still present.